requirements.txt
streamlit
langchain-openai
langchain-community
langchain-huggingface
supabase
faiss-cpu
pypdf
python-dotenv
pandas
sqlalchemy
reportlab
